+++ src/IO/cuda-md5/md5Opt.h
@@ -15,7 +15,7 @@ extern "C" {
 #endif
 typedef unsigned int MD5_u32plus;
 int FTI_destroyMD5();
-int FTI_initMD5(long cSize, long tempSize, FTIT_configuration *FTI_Conf);
+int FTI_initMD5(int32_t cSize, int32_t tempSize, FTIT_configuration *FTI_Conf);
 int FTI_MD5CPU(FTIT_dataset *data);
 int FTI_MD5GPU(FTIT_dataset *data);
 int FTI_SyncMD5();
