+++ src/conf.c
@@ -167,12 +167,12 @@ int FTI_ReadConf(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
     FTI_Conf->dcpInfoPosix.StackSize = (int)iniparser_getint(ini,
      "Basic:dcp_stack_size", 5);
 
-    long long maxVarId = (long long)iniparser_getlint(ini, "Basic:max_var_id",
-     (long long)FTI_DEFAULT_MAX_VAR_ID);
-    if (maxVarId > (long long)FTI_LIMIT_MAX_VAR_ID) {
+    int64_t maxVarId = (int64_t)iniparser_getlint(ini, "Basic:max_var_id",
+     (int64_t)FTI_DEFAULT_MAX_VAR_ID);
+    if (maxVarId > (int64_t)FTI_LIMIT_MAX_VAR_ID) {
         char err[FTI_BUFS];
         snprintf(err, FTI_BUFS, "Value of 'Basic:max_var_id' cannot be higher"
-        " than 'FTI_LIMIT_MAX_VAR_ID' ('%lld > %d')", maxVarId,
+        " than 'FTI_LIMIT_MAX_VAR_ID' ('%ld > %d')", maxVarId,
          FTI_LIMIT_MAX_VAR_ID);
         FTI_Print(err, FTI_WARN);
         FTI_Conf->maxVarId = FTI_DEFAULT_MAX_VAR_ID;
